pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Julia Bengert <yubengert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"
drf-spectacular = "^0.28.0"
pytelegrambotapi = "^4.26.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
